Reset to single threaded and avoid cellId overflow

+++ openrouteservice/src/main/java/heigit/ors/partitioning/FlowEdgeData.java
@@ -1,12 +1,12 @@
 package heigit.ors.partitioning;
 
 public class FlowEdgeData {
-    public short flow = 0;
-    public short capacity = 0;
+    public int flow = 0;
+    public int capacity = 0;
     public int inverse = -1;
     public boolean active = false;
 
-    public FlowEdgeData(short flow, short capacity, int inverse, boolean active) {
+    public FlowEdgeData(int flow, int capacity, int inverse, boolean active) {
         this.flow = flow;
         this.capacity = capacity;
         this.inverse = inverse;
